team-api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "craic-team-api"
version = "0.1.0"
description = "CRAIC team knowledge store API"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2.0",
    "pyjwt>=2.0",
    "bcrypt>=4.0",
]

[project.scripts]
craic-team-api = "team_api.app:main"

[dependency-groups]
dev = [
    { include-group = "lint" },
    { include-group = "tests" },
]

lint = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.4",
]

tests = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx",
]

[tool.hatch.build.targets.wheel]
packages = ["team_api"]

[tool.ruff.lint]
ignore = [
    # FastAPI uses Depends() in function default arguments; this is intentional and idiomatic.
    "B008",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
